CMS Tutorial - Authentication #1
Installing Authentication Plugin
https://book.cakephp.org/4/en/tutorials-and-examples/cms/authentication.html#installing-authentication-plugin

# composer require "cakephp/authentication:^2.0"
# bin/cake bake all users

Adding Password Hashing
https://book.cakephp.org/4/en/tutorials-and-examples/cms/authentication.html#adding-password-hashing

In CakePHP, we use different classes to operate on collections of records and single records. Methods that operate on the collection of entities are put in the Table class, while features belonging to a single record are put on the Entity class.

CakePHP hashes passwords with bcrypt by default. We recommend bcrypt for all new applications to keep your security standards high. This is the recommended password hash algorithm for PHP.
Note: bcrypt will generate a different hash even for the same password saved twice.